Add kata1 package files. Restructure location of unit tests and main .py file
