+++ kata1_dictionary/README.md
@@ -1,0 +1,13 @@
+# KATA \#1 Create a Dictionary class in Python
+
+In this kata, your job is to create a class Dictionary which you can add words to and their entries. Example:
+
+>>> d = Dictionary()
+
+>>> d.newentry('Apple', 'A fruit that grows on trees')
+
+>>> print(d.look('Apple'))
+A fruit that grows on trees
+
+>>> print(d.look('Banana'))
+Can't find entry for Banana
