+++ kata1_dictionary/pyproject.toml
@@ -1,0 +1,6 @@
+[build-system]
+requires = ['setuptools>=42']
+build-backend = 'setuptools.build_meta'
+
+[tool.setuptools]
+package-dir = {"" = "src"}
