+++ kata1_dictionary/src/kata1/kata1_dictionary.py
